+++ core/src/main/java/gay/oss/gatos/core/collection/FlowCollection.java
@@ -1,4 +1,4 @@
-package gay.oss.gatos.core.collections;
+package gay.oss.gatos.core.collection;
 
 import gay.oss.gatos.core.models.Flow;
